+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 ### Changed
 * Adjust uniqueness check on ps_component/purl (`OSIDB-4667`)
 
+### Changed
+* Updated aegis feedback method (`AEGIS-203`)
+
 
 ## [2025.11.4]
 ### Added